Move API calls to `/features/api`, fix tests
#47
